Database Layout Pre-Update
Since I am going to change the database layout made this commit to role back and see the changes that might have broke something. So, with the explanation aside, heres what this contains from the last commit. Authentication works with email, and email verification is required if signing in with email (although the verification email is not sent anywhere, since we have no email client), the account settings tab is also synced and the currency and stuff.