Patched a few things, and removed references to SimpleJSON
So. The DB was mean was it... kept crashing the server if it was set to disabled. Well now it won't do that because the instance will be defined but it won't do anything unless it is enabled. And yes, now the config uses System.Text.Json namespace instead of SimpleJSON.